Drop unused library module
It's not necessary to have a library module to demonstrate a working
version of using the Shadow plugin with AAR dependencies. It's
sufficient to simultaneously try to instal two different versions of the
same plugin within the same Android app.

Gradle does not allow for multiple versions of the same dependencies so
if a single app can build with two versions of the same dependency, that
means that the Shadow plugin has successfully moved one version of the
dependency to a different namespace to avoid collissions.

So, let's delete this library to reduce complexity